+++ app/models/criterion.rb
@@ -16,6 +16,9 @@ class Criterion < ApplicationRecord
            through: :criteria_assignment_files_joins
   accepts_nested_attributes_for :criteria_assignment_files_joins, allow_destroy: true
 
+  validate :results_unreleased?
+  validate :visible?
+
   self.abstract_class = true
 
   # Assigns a random TA from a list of TAs specified by +ta_ids+ to each
@@ -161,6 +164,13 @@ class Criterion < ApplicationRecord
       )
     end
     a.assignment_stat.refresh_grade_distribution
+  end
+
+  def results_unreleased?
+    return true if self.marks.joins(:result).where('results.released_to_students' => true).empty?
+
+    errors.add(:base, 'Cannot update criterion once results are released.')
+    false
   end
 
   private