+++ spec/factories/levels.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :level do
+    name { Faker::Lorem.word }
+    description { Faker::Lorem.word }
+    rubric_criterion
+  end
+end